Refactor console application (#603)
